+++ .travis.yml
@@ -4,7 +4,7 @@ branches:
   - master
   - devops
 python:
-  - "3.5"
+  - "3.7"
   - "3.8"
 # command to install dependencies
 before-install:
