+++ frontend/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "Music Player",
+  "name": "Music Player React App",
   "icons": [
     {
       "src": "favicon.ico",
