+++ frontend/src/App.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import NotFound from './NotFound';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
+  ReactDOM.render(<NotFound />, div);
   ReactDOM.unmountComponentAtNode(div);
 });
