+++ frontend/src/ReduxStore.js
@@ -1,0 +1,6 @@
+import configureStore from './store/store'
+import { LogoutState } from './store/reducers/UserUpdater'
+
+const store = configureStore(LogoutState)
+
+export default store;
